% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nondifferentiation.R
\name{calc_nondifferentiation}
\alias{calc_nondifferentiation}
\title{Calculation of nondifferentiation over multiple item packages}
\usage{
calc_nondifferentiation(
  df = df,
  items,
  new_var_name,
  proportion_allowed_missings = 0.1
)
}
\arguments{
\item{df}{The dataframe of interest which includes the columns of interest}

\item{items}{A list of vectors of colnames()}

\item{new_var_name}{A vector of new names for the variables to be generated}

\item{proportion_allowed_missings}{the proportion of missings per row and item package. CAVE! missings means values < 0 only. Default: no value is being calculated if > 10\% of values are missing in an item package. In that case, NA is displayed}
}
\value{
A dataframe will be printed to the console
}
\description{
Function calculates the mean of square root differences of all pairs of values in an item package per person. This is then standardized using the minimum and maximum values in the dataset.The values range from 0 (least differentiation) to 1 (maximum differentiation). 0 in many cases means, the row contains the same value for all items of the considered item package. The procedure follows the description of Kim et al., 2019. The function is applicable to a list of several item packages. A vector of names needs to be defined. The function automatically sets the value of a respondent NA if more than 10\% of the items' values are < 0. The proportion is adjustable. The function does not have a rule included for dealing with NAs, yet.
}
\examples{
library(data.table)
library(dplyr)
set.seed(1313)
df <- data.table(
  first01 = c(4, -2, 4, 5, 6, 2, -1, -1, -9, 4, 5, 1, 3, 4, 3, 4, 5, -1, -1, 5, 2, 3, 4, 3, 2, 2, -1, -9, 2, 1),
  first02 = c(-1, -9, sample(1:5, size = 27, replace = TRUE), -1),
  first03 = c(sample(1:5, size = 30, replace = TRUE)),
  first04 = c(sample(1:5, size = 30, replace = TRUE)),
  first05 = c(sample(1:5, size = 30, replace = TRUE)),
  first06 = c(sample(1:5, size = 30, replace = TRUE)),
  first07 = c(sample(1:5, size = 30, replace = TRUE)),
  first08 = c(sample(1:5, size = 30, replace = TRUE)),
  second01 = c(sample(1:11, size = 30, replace = TRUE)),
  second02 = c(sample(1:11, size = 30, replace = TRUE)),
  second03 = c(sample(1:11, size = 30, replace = TRUE)),
  second04 = c(sample(1:11, size = 30, replace = TRUE)),
  second05 = c(sample(1:11, size = 30, replace = TRUE)),
  second06 = c(sample(1:11, size = 30, replace = TRUE)),
  second07 = c(sample(1:11, size = 30, replace = TRUE)),
  second08 = c(sample(1:11, size = 30, replace = TRUE)),
  third01 = c(sample(1:5, size = 30, replace = TRUE)),
  third02 = c(sample(1:5, size = 30, replace = TRUE)),
  third03 = c(sample(1:5, size = 30, replace = TRUE)),
  third04 = c(sample(1:5, size = 30, replace = TRUE)),
  third05 = c(sample(1:5, size = 30, replace = TRUE)),
  third06 = c(sample(1:5, size = 30, replace = TRUE)),
  fourth = c(sample(1:2, size = 30, replace = TRUE)),
  fifth = c(sample(1:11, size = 30, replace = TRUE)))

  df[sample(1:30, size = 4), ] <- -10

  items_first <- df \%>\% select(contains("first")) \%>\% colnames()
  items_second <- df \%>\% select(contains("second")) \%>\% colnames()
  items_third <- df \%>\% select(contains("third")) \%>\% colnames()

  items_list <- list(items_first, items_second, items_third)
  names_list <- c("firstND", "secondND", "thirdND")

  for (i in seq_along(items_list)) {
  df <- calc_nondifferentiation(df, items_list[[i]], names_list[i])
  }

  print(df)









}
\references{
Kim, Y., Dykema, J., Stevenson, J., Black, P., & Moberg, D. P. (2019). Straightlining: Overview of Measurement, Comparison of Indicators, and Effects in Mail–Web Mixed-Mode Surveys. Social Science Computer Review, 37(2), 214–233. https://doi.org/10.1177/0894439317752406
}
\author{
Julia Witton
}
